Add Events Seeder
Events All and Create Visualization

+++ ASP.Net Core/Eventures/Eventures/Data/Models/Event.cs
@@ -21,6 +21,6 @@
 
         public int Tickets { get; set; }
 
-        public decimal TicketPrice { get; set; }
+        public double TicketPrice { get; set; }
     }
 }
