+++ ASP.Net Core/Eventures/Eventures/ViewModels/EventInputModel.cs
@@ -14,6 +14,6 @@ namespace Eventures.ViewModels
 
         public int Tickets { get; set; }
 
-        public decimal TicketPrice { get; set; }
+        public double TicketPrice { get; set; }
     }
 }
